+++ lib/diamond/src/lib.rs
@@ -1,7 +1,5 @@
 mod c_api;
 mod geometry;
 mod mesh;
-#[cfg(test)]
-mod tests;
 
 pub use c_api::*;
