+++ lib/diamond/test-data/asymetric-collapse-collision-mesh.json
