+++ lib/diamond/test-data/asymetric-collapse-input.json
