+++ lib/diamond/test-data/large-collision-mesh.json
