+++ lib/diamond/test-data/large-input.json
