+++ lib/diamond/test-data/overlap-collision-mesh.json
