+++ lib/diamond/test-data/small-collision-mesh.json
