+++ lib/diamond/test-data/small-input.json
