+++ lib/diamond/test-data/trivial-collision-mesh.json
