+++ lib/diamond/test-data/trivial-input.json
